feat: send a message when failing to recalculate disguise
